add helper functions for debit routing
